colorbalancergb: add ColorWheel control
Although the sliders are a convenient way of setting the various parameters,
they have the caveat of making it hard to visualize common color theory patterns.
One has to either use a 3rd party software with a color wheel or resort to doing
mental maths to compute opposite or adjacent hue values.

For this reason, this change adds a simple color wheel control which allows to
visualize which hue/chroma the shadows/midtones/highlights are being set to.

It can serve purely for visualization purposes or can be interacted with as an
alternate way of changing hues / chromas.

As the chroma is rarely set to a very high value and more often to low values,
the color wheel uses a non linear scaling of the chroma values so as to make
it easier to set values in the usual target range.

- Maintaining Shift while dragging make changing hue in steps of 10 degrees and
  hues in step of 0.1 on the non linear scale.

- Maintaining Control while dragging locks the chroma so that only hues can be
  changed

- Mouse Scroll will increase/decrease hue by 10 of closest cursor on the color
  wheel